+++ ninshiki_py/ninshiki_py_detector.py
@@ -22,8 +22,8 @@ import argparse
 import rclpy
 from rclpy.node import Node
 
-from ninshiki_yolo.detector.detection import Detection
-from ninshiki_yolo.node.ninshiki_yolo_node import NinshikiYoloNode
+from ninshiki_py.detector.yolo import Yolo
+from ninshiki_py.node.ninshiki_py_node import NinshikiPyNode
 
 
 def main(args=None):
@@ -37,10 +37,10 @@ def main(args=None):
                         type=int, choices=[0, 1], default=0)
     arg = parser.parse_args()
 
-    node = Node("ninshiki_yolo")
+    node = Node("ninshiki_py")
 
-    detection = Detection(gpu=arg.GPU, myriad=arg.MYRIAD)
-    detector_node = NinshikiYoloNode(node, arg.topic, detection)
+    detection = Yolo(gpu=arg.GPU, myriad=arg.MYRIAD)
+    detector_node = NinshikiPyNode(node, arg.topic, detection)
 
     rclpy.spin(detector_node.node)
