+++ ninshiki_py/utils/draw_detection_result.py
