+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>ninshiki_yolo</name>
+  <name>ninshiki_py</name>
   <version>0.1.0</version>
   <description>An object detection package with YOLO</description>
   <maintainer email="nathanael.hutama24@gmail.com">ichiro</maintainer>
