+++ setup.cfg
@@ -1,4 +1,4 @@
 [develop]
-script-dir=$base/lib/ninshiki_yolo
+script-dir=$base/lib/ninshiki_py
 [install]
-install-scripts=$base/lib/ninshiki_yolo
+install-scripts=$base/lib/ninshiki_py
